pipeline/pyproject.toml
[tool.poetry]
name = "pipeline"
version = "0.1.0"
description = ""
authors = ["Filip Trplan <filip@trplan.si>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
typer = "0.9.0"
music21 = "^9.1.0"
soundfile = "^0.12.1"
typer-config = {extras = ["yaml"], version = "^1.4.0"}
elasticsearch = "^8.12.1"
python-dotenv = "^1.0.1"
essentia = "^2.1b6.dev1110"
pesto-pitch = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
syrupy = "^4.6.1"
ruff = "^0.3.0"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
